Added a function to get positions and 2 classes
Added a piece class and a position class
Position class will make it easier for us to handle board configuration
using a 2D grid
Added a getPositions() function that get will eventually get all of the
available positions of a piece
The function is invoked when the user clicks on a button that is not
empty. Right now
only the "P" from leftmost bottom and "Kn" from leftmost bottom calls
the function, the available positions will be displayed on console.